+++ packages/db/drizzle/0006_cooing_moonstone.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `relay_index` ADD `sender` text;